Remove unnecessary test in DataValuesTestBase
